Increase the data block size limit and improve __microfs_read() caching.
The previous block size limit was probably put in place due to an early
misunderstanding of how zlib works. The new limit is more or less arbitrarily
chosen - to limit blocks to 1MB seems about right.

__microfs_read() will now search all available read buffers for cache hits
before reading data from the image. This should allow microfs to handle
small files a little more efficiently.

+++ hostprog_microfsmki.c
@@ -883,7 +883,7 @@ static struct imgspec* create_imgspec(int argc, char* argv[])
 	/* The block size should now be correctly set, which means
 	 * that the block left shift can be calculated.
 	 */
-	__u16 blksz = spec->sp_blksz;
+	size_t blksz = spec->sp_blksz;
 	while ((blksz >>= 1) > 0)
 		spec->sp_blkshift++;
